+++ tests/unit/models/watcher-test.js
@@ -27,6 +27,23 @@ describe('Watcher', function() {
     });
   });
 
+  describe('watcher strategy selection', function() {
+    it('selects the events-based watcher by default', function () {
+      subject.options = null;
+      assert.equal(subject.selectWatcher(), require('broccoli-sane-watcher'));
+    });
+
+    it('selects the events-based watcher when given events watcher option', function () {
+      subject.options = { watcher: 'events' };
+      assert.equal(subject.selectWatcher(), require('broccoli-sane-watcher'));
+    });
+
+    it('selects the polling watcher when given polling watcher option', function () {
+      subject.options = { watcher: 'polling' };
+      assert.equal(subject.selectWatcher(), require('broccoli/lib/watcher'));
+    });
+  });
+
   describe('watcher:change', function() {
     beforeEach(function() {
       watcher.emit('change', {